requirements.txt
discord.py==2.3.2
matplotlib==3.9.0
numpy==2.0.0
requests==2.32.3
python-dateutil==2.9.0.post0
Pillow==10.4.0
python-dotenv
